+++ kolibri/plugins/coach/assets/src/views/assignments/AssignmentDeleteModal.vue
