+++ kolibri/plugins/coach/assets/src/views/assignments/StatusIcon.vue
